# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Protean API
  description: |
    REST and WebSocket API for the Protean Device Farm.
    
    ### Authentication
    The API supports two authentication mechanisms:
    1. **JWT Bearer Token**: Sent in the `Authorization` header as `Bearer <token>`.
    2. **API Key**: Sent in the `X-API-Key` header for programmatic or machine-to-machine integrations.
    
    In local development, authentication can be bypassed if the environment variable `BYPASS_AUTH_IN_DEV` is set to `true`.
  version: 1.0.0
servers:
  - url: http://localhost:9002
    description: Local Coordinator Service (HTTP & WebSockets)
paths:
...
  /api/v1/devices:
    get:
      summary: List Connected Devices
      description: Returns a list of mobile devices connected to the device farm. If authenticated, the list is scoped according to the user's group permissions (administrators see all devices).
      tags:
        - Devices
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: A list of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '500':
          description: Failed to query database
  /api/v1/devices/ws:
    get:
      summary: Device Stream WebSocket
      description: Establishes a persistent WebSocket connection to receive real-time device updates (e.g., state updates, claims, releases).
      tags:
        - Devices
      parameters:
        - name: token
          in: query
          required: false
          description: Optional auth token if not provided in the standard connection headers
          schema:
            type: string
      responses:
        '101':
          description: Switching Protocols to WebSocket.
          headers:
            Upgrade:
              schema:
                type: string
                example: websocket
            Connection:
              schema:
                type: string
                example: Upgrade
...
  /api/v1/automation/scripts:
    get:
      summary: List Automation Scripts
      description: Returns all saved YAML automation scripts.
      tags:
        - Automation
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of automation scripts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScriptDB'
        '500':
          description: Database query failed
    post:
      summary: Save/Update Automation Script
      description: Validates and saves a YAML-defined automation script. If ID is provided, it updates the existing script; otherwise, a new UUID is generated.
      tags:
        - Automation
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - content
              properties:
                id:
                  type: string
                  format: uuid
                  description: Optional ID. If supplied, updates the existing script.
                  example: 44444444-4444-4444-4444-444444444444
                name:
                  type: string
                  description: Name/Label of the script
                  example: Open Youtube and Play Video
                content:
                  type: string
                  description: YAML DSL Script content matching the automation framework schema.
                  example: |
                    variables:
                      step_delay_ms: "500"
                    steps:
                      - launch: { package: "com.google.android.youtube" }
                      - wait: { condition: "visible", xpath: "//node[@content-desc='Search']", timeoutMs: 5000 }
                      - click: { xpath: "//node[@content-desc='Search']" }
                      - input: { text: "Nature 4k video" }
      responses:
        '200':
          description: Script saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  id:
                    type: string
                    format: uuid
                    example: 44444444-4444-4444-4444-444444444444
        '400':
          description: Invalid request body or malformed YAML DSL content
        '500':
          description: Database save failed
...
  /api/v1/automation/run:
    post:
      summary: Execute Script Parallel
      description: Triggers the execution of a saved script in parallel across one or more target devices.
      tags:
        - Automation
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - script_id
                - device_serials
              properties:
                script_id:
                  type: string
                  format: uuid
                  example: 44444444-4444-4444-4444-444444444444
                device_serials:
                  type: array
                  items:
                    type: string
                  example: ["RCLBYQWFES4ZLR4", "00008020-001A4D3E1402002E"]
      responses:
        '200':
          description: Execution completed, returns the run outcome summary for each device
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        serial:
                          type: string
                          example: RCLBYQWFES4ZLR4
                        report_id:
                          type: string
                          format: uuid
                          example: 55555555-5555-5555-5555-555555555555
                        success:
                          type: boolean
                          example: true
                        duration_ms:
                          type: integer
                          example: 12450
                        error:
                          type: string
                          description: Error message if step failed
                          example: "click failed: element not found"
        '400':
          description: Invalid body, missing fields, or empty target device array
        '404':
          description: Script not found
        '500':
          description: Execution runner or database error
  /api/v1/automation/reports:
    get:
      summary: List Automation Execution Reports
      description: Returns metadata of all script run historical reports.
      tags:
        - Automation
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of execution reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportDB'
        '500':
          description: Database query failed
...
  /api/v1/admin/users:
    get:
      summary: List Users (Admin Only)
      description: Lists all registered users and their roles. Requires administrator privileges.
      tags:
        - Admin Management
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: Forbidden (Administrator privileges required)
        '500':
          description: Database query failed
    delete:
      summary: Delete User (Admin Only)
      description: Permanently deletes a user by ID. Requires administrator privileges.
      tags:
        - Admin Management
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: query
          required: true
          description: The UUID of the user to delete
          schema:
            type: string
            format: uuid
            example: 00000000-0000-0000-0000-000000000000
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: Missing ID parameter
        '403':
          description: Forbidden (Administrator privileges required)
        '500':
          description: Database deletion failed
  /api/v1/admin/groups:
    get:
      summary: List Groups (Admin Only)
      description: Lists all logical tenant groups. Requires administrator privileges.
      tags:
        - Admin Management
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of tenant groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '403':
          description: Forbidden
        '500':
          description: Database query failed
    post:
      summary: Create Tenant Group (Admin Only)
      description: Creates a new tenant/device group. Requires administrator privileges.
      tags:
        - Admin Management
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Group name (must be unique)
                  example: Quality Assurance
                description:
                  type: string
                  description: Short summary of group scope
                  example: Group for manual and automated QA test runs
                expires_at:
                  type: string
                  format: date-time
                  description: Optional ISO-8601 expiry timestamp for temporal groups
                  example: "2026-12-31T23:59:59Z"
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid request payload or missing name
        '403':
          description: Forbidden
        '500':
          description: Database insertion failed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter the JSON Web Token (JWT) received after login.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Programmatic API Key token generated inside user auth keys panel.
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 00000000-0000-0000-0000-000000000000
        email:
          type: string
          format: email
          example: admin@domain.com
        role:
          type: string
          enum: [admin, group_admin, user, viewer]
          example: admin
        auth_provider:
          type: string
          example: local
        created_at:
          type: string
          format: date-time
          example: "2026-07-05T07:36:54Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-07-05T07:36:54Z"
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 11111111-1111-1111-1111-111111111111
        name:
          type: string
          example: Public
        description:
          type: string
          example: Default public group
        created_at:
          type: string
          format: date-time
          example: "2026-07-05T07:36:54Z"
        expires_at:
          type: string
          format: date-time
          nullable: true
          example: null
...
    Device:
      type: object
      properties:
        serial:
          type: string
          example: RCLBYQWFES4ZLR4
        provider_id:
          type: string
          description: Host IP of the provider managing this device.
          example: 192.168.1.100
        model:
          type: string
          example: realme RMX3085
        manufacturer:
          type: string
          example: realme
        platform:
          type: string
          enum: [android, ios]
          example: android
        os_version:
          type: string
          example: "12"
        android:
          type: string
          description: Legacy version compatibility field (matches os_version).
          example: "12"
        sdk:
          type: integer
          example: 31
        abi:
          type: string
          example: arm64-v8a
        ram_mb:
          type: integer
          format: int64
          example: 8192
        storage_mb:
          type: integer
          format: int64
          example: 128000
        display:
          type: string
          example: 1080x2400 @ 405dpi
        battery:
          type: integer
          example: 92
        wifi_ssid:
          type: string
          example: Corporate-Wifi
        ip:
          type: string
          example: 192.168.1.115
        status:
          type: string
          enum: [idle, claimed, busy, offline]
          example: idle
        stream_port:
          type: integer
          description: Allocated MJPEG stream port. 0 if not claimed.
          example: 0
        connected_at:
          type: string
          format: date-time
          example: "2026-07-09T23:26:00Z"
        file_system:
          type: object
          description: File hierarchy JSON data (if cached/browsed).
          nullable: true
          example: null
        installed_browsers:
          type: array
          items:
            type: string
          description: Installed web browsers detected on the device.
          nullable: true
          example: ["com.android.chrome"]
    ScriptDB:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 44444444-4444-4444-4444-444444444444
        name:
          type: string
          example: Launch Chrome
        content:
          type: string
          description: Raw YAML DSL script content.
          example: |
            variables:
              step_delay_ms: "500"
            steps:
              - launch: { package: "com.android.chrome" }
        created_at:
          type: string
          format: date-time
          example: "2026-07-06T15:30:00Z"
    ReportDB:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 55555555-5555-5555-5555-555555555555
        script_id:
          type: string
          format: uuid
          example: 44444444-4444-4444-4444-444444444444
        serial:
          type: string
          example: RCLBYQWFES4ZLR4
        success:
          type: boolean
          example: true
        start_time:
          type: string
          format: date-time
          example: "2026-07-06T15:31:00Z"
        end_time:
          type: string
          format: date-time
          example: "2026-07-06T15:31:12Z"
...